+++ oo_patterns/behavioral/observer/sub_space_publishers.py
@@ -7,7 +7,10 @@ import logging
 import typing as t
 from collections import defaultdict
 
-from .interfaces import AsyncSubSpacePublisherInterface, SubSpacePublisherInterface
+from .interfaces import (
+    AsyncSubSpacePublisherInterface,
+    SubSpacePublisherInterface,
+)
 from .publishers import AsyncPublisher, Publisher
 
 if t.TYPE_CHECKING:
@@ -19,7 +22,8 @@ EventContextTypeVar = t.TypeVar("EventContextTypeVar")
 
 
 class BaseSubSpacePublisher(
-    SubSpacePublisherInterface[EventContextTypeVar], t.Generic[EventContextTypeVar]
+    SubSpacePublisherInterface[EventContextTypeVar],
+    t.Generic[EventContextTypeVar],
 ):
     def __init__(self):
         self._sub_spaces_publishers_map: dict[
@@ -39,7 +43,9 @@ class BaseSubSpacePublisher(
 
         self._sub_spaces_publishers_map[sub_space].remove_subscribers(*args)
 
-    def notify_subscribers(self, context: EventContextTypeVar) -> list[Exception]:
+    def notify_subscribers(
+        self, context: EventContextTypeVar
+    ) -> list[Exception]:
         sub_space = self.get_event_context_sub_space(context=context)
         if sub_space not in self._sub_spaces_publishers_map:
             return []
@@ -54,7 +60,8 @@ class BaseSubSpacePublisher(
 
 
 class BaseAsyncSubSpacePublisher(
-    AsyncSubSpacePublisherInterface[EventContextTypeVar], t.Generic[EventContextTypeVar]
+    AsyncSubSpacePublisherInterface[EventContextTypeVar],
+    t.Generic[EventContextTypeVar],
 ):
 
     def __init__(self):
@@ -79,13 +86,15 @@ class BaseAsyncSubSpacePublisher(
 
         self._sub_spaces_publishers_map[sub_space].remove_subscribers(*args)
 
-    async def notify_subscribers(self, context: EventContextTypeVar) -> list[Exception]:
+    async def notify_subscribers(
+        self, context: EventContextTypeVar
+    ) -> list[Exception]:
         sub_space = self.get_event_context_sub_space(context=context)
         if sub_space not in self._sub_spaces_publishers_map:
             return []
-        return await self._sub_spaces_publishers_map[sub_space].notify_subscribers(
-            context=context
-        )
+        return await self._sub_spaces_publishers_map[
+            sub_space
+        ].notify_subscribers(context=context)
 
     def remove_all_subscribers(self):
         for publisher in self._sub_spaces_publishers_map.values():