+++ oo_patterns/behavioral/state/interfaces.py
@@ -10,7 +10,9 @@ if t.TYPE_CHECKING:
 
 
 StateTypeVar = t.TypeVar("StateTypeVar", bound="StateInterface")
-StateContextTypeVar = t.TypeVar("StateContextTypeVar", bound="StateContextInterface")
+StateContextTypeVar = t.TypeVar(
+    "StateContextTypeVar", bound="StateContextInterface"
+)
 
 
 class StateInterface(t.Generic[StateContextTypeVar]):
@@ -19,11 +21,9 @@ class StateInterface(t.Generic[StateContextTypeVar]):
     def set_context(self, context: StateContextTypeVar):
         raise NotImplementedError(f"{self.__class__}.set_context")
 
-    def init_state(self):
-        raise NotImplementedError(f"{self.__class__}.init_state")
-
 
 class StateContextInterface(t.Generic[StateTypeVar]):
+    proxy: t.Self
 
     def set_state(self, state: StateTypeVar):
         raise NotImplementedError(f"{self.__class__}.set_state")