+++ tests/behavioral/state/test_base.py
@@ -1,11 +1,15 @@
+import typing as t
+import weakref
+from functools import cached_property
 from unittest import TestCase
 
 from oo_patterns.behavioral.state import BaseState, BaseStateContext
 
+if t.TYPE_CHECKING:
+    ...
+
 
 class BaseTestState(BaseState["TestStateContext"]):
-
-    def init_state(self): ...
 
     def transition_to_first(self):
         raise NotImplementedError
@@ -28,6 +32,10 @@ class SecondTestState(BaseTestState):
 class TestStateContext(BaseStateContext["BaseTestState"]):
     def __init__(self):
         self.set_state(state=FirstTestState())
+
+    @cached_property
+    def proxy(self) -> t.Self:  # type: ignore[override]
+        return weakref.proxy(self)
 
     def transition_to_first(self):
         self.state.transition_to_first()